Add ability to disable and enable linting within a file (#91)
This PR does two things:

1. It adds `rapids-pre-commit-hooks: (disable|enable)(-next-line)?`
directives
2. It improves the testing framework by creating a domain-specific
language that visually spells out where the expected locations in a
piece of text are

The motivation for the directives is to allow us to not complain about
hard-coded versions in `pyproject.toml` dependencies and remove the
f-string specifiers from `test_copyright.py`.

The motivation for the changes to the testing framework is obvious:
manually calculating where all the locations in a piece of text are is a
horrible pain, and is subject to breakage if the text changes. This new
framework allows us to easily read and write the location values. It
also allows us to very easily construct lists and objects from these
locations, thus negating the need to pass an "expected result" parameter
to the test in some cases.